feat: Add buildkit container with exposed port (#56)
* feat: Add buildkit container with exposed port

This is useful for use in GitLab CI/CD, where services have to undergo a
health check. Since the default buildkit image does not `EXPOSE` a port,
the healthcheck will fail. This image is intended as a replacement: all
it does is add the `EXPOSE` instruction and sets a default `CMD` that
works in GitLab CI/CD.

* docs: Add buildkit image to the readme

* chore(renovate): Automatically update buildkit digests